==> packages/wmb/wmb-0.0.19.tar.gz/wmb-0.0.19/analysis/integration/mc_3c/Template/04.integration.ipynb <==
from ALLCools.clustering import *
from wmb import brain, cemba, aibs
import numpy as np
import pandas as pd
import anndata

import matplotlib.pyplot as plt
from ALLCools.plot import *

from ALLCools.integration.seurat_class import SeuratIntegration

dataset = 'cemba_3c'
categorical_key = ['L1_annot', 'L2', 'DissectionRegion']

## Input LSI before integration

mc_adata = anndata.read_h5ad('mc_pca.h5ad')
m3c_adata = anndata.read_h5ad('m3c_pca.h5ad')

adata_list = [mc_adata, m3c_adata]

### Init empty adata_merge

from scipy.sparse import csr_matrix

cells = sum([a.shape[0] for a in adata_list])
features = adata_list[0].shape[1]

adata_merge = anndata.AnnData(X=csr_matrix(([], ([], [])),
                                           shape=(cells, features), dtype='float32'),
                              obs=pd.concat([a.obs for a in adata_list]),
                              var=adata_list[0].var)

mc_annot = cemba.get_mc_annot()
m3c_annot = cemba.get_m3c_annot()

for key in categorical_key:
    adata_merge.obs[key] = pd.concat(
        [mc_annot[key].to_pandas(), 
         m3c_annot[key].to_pandas()]
    )

for adata in adata_list:
    for key in categorical_key:
        adata.obs[key] = adata_merge.obs[key]

adata_list

adata_merge

## Integration and transform LSI

integrator = SeuratIntegration()

anchor = integrator.find_anchor(adata_list,
                                k_local=None,
                                key_local='X_pca',
                                k_anchor=5,
                                key_anchor='X',
                                dim_red='cca',
                                max_cc_cells=100000,
                                k_score=30,
                                k_filter=None,
                                scale1=False,
                                scale2=False,
                                n_components=50,
                                n_features=200,
                                alignments=[[[0], [1]]])

corrected = integrator.integrate(key_correct='X_pca',
                                 row_normalize=True,
                                 n_components=30,
                                 k_weight=100,
                                 sd=1,
                                 alignments=[[[0], [1]]])

adata_merge.obsm['X_pca_integrate'] = np.concatenate(corrected)

## Label transfer

transfer_results = integrator.label_transfer(
    ref=[0],
    qry=[1],
    categorical_key=categorical_key,
    key_dist='X_pca'
)

for k, v in transfer_results.items():
    v.to_hdf(f'{k}_transfer.hdf', key='data')

integrator.save_transfer_results_to_adata(adata_merge, transfer_results)

## Save

adata_merge.write_h5ad('final.h5ad')

adata_merge

integrator.save('integration')
